requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
sqlparse==0.5.5
